@@ -1,6 +1,6 @@
 {
   "name": "mki-formsy-react",
-  "version": "0.22.0",
+  "version": "0.23.0",
   "description": "A form input builder and validator for React JS",
   "repository": {
     "type": "git",
@@ -27,26 +27,31 @@
     "react-component"
   ],
   "dependencies": {
-    "form-data-to-object": "^0.2.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.10.1",
-    "babel-core": "^6.10.4",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-2": "^6.11.0",
-    "jsdom": "^9.4.1",
-    "nodeunit": "^0.9.1",
-    "react": "^15.2.0",
-    "react-addons-pure-render-mixin": "^15.2.0",
-    "react-addons-test-utils": "^15.2.0",
-    "react-dom": "^15.2.0",
-    "sinon": "^1.17.4",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1"
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
+    "jsdom": "^11.0.0",
+    "nodeunit": "^0.11.0",
+    "react": "^15.5.4",
+    "react-addons-pure-render-mixin": "^15.5.2",
+    "react-dom": "^15.5.4",
+    "sinon": "^1.17.7",
+    "webpack": "^2.6.1",
+    "webpack-dev-server": "^2.4.5",
+    "create-react-class": "^15.5.0",
+    "form-data-to-object": "^0.2.0",
+    "prop-types": "^15.5.0",
+    "react": "^15.0.0 || ^16.0.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
+    "create-react-class": "^15.5.0",
+    "form-data-to-object": "^0.2.0",
+    "prop-types": "^15.5.0",
+    "react": "^15.0.0 || ^16.0.0"
   }
 }
