@@ -1,5 +1,6 @@
 var utils = require('./utils.js');
 var React = global.React || require('react');
+var PropTypes = require('prop-types');
 
 var convertValidationsToObject = function (validations) {
 
@@ -48,7 +49,7 @@ module.exports = {
     return this.props.value != null ? this.props.value : this.props.defaultValue;
   },
   contextTypes: {
-    formsy: React.PropTypes.object // What about required?
+    formsy: PropTypes.object // What about required?
   },
   getDefaultProps: function () {
     return {
