@@ -1,5 +1,5 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 import PureRenderMixin from 'react-addons-pure-render-mixin';
 import sinon from 'sinon';
 import ReactDOM from 'react-dom';
@@ -217,16 +217,15 @@ export default {
   },
 
   'should allow an undefined value to be updated to a value': function (test) {
+    class TestForm extends React.Component {
+      state = {value: undefined};
 
-    const TestForm = React.createClass({
-      getInitialState() {
-        return {value: undefined};
-      },
-      changeValue() {
+      changeValue = () => {
         this.setState({
           value: 'foo'
         });
-      },
+      };
+
       render() {
         return (
           <Formsy.Form url="/users">
@@ -234,7 +233,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     form.changeValue();
@@ -243,12 +243,10 @@ export default {
       test.equal(input.value, 'foo');
       test.done();
     });
-
   },
 
   'should be able to test a values validity': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -256,19 +254,18 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const input = TestUtils.findRenderedComponentWithType(form, TestInput);
     test.equal(input.isValidValue('foo@bar.com'), true);
     test.equal(input.isValidValue('foo@bar'), false);
     test.done();
-
   },
 
   'should be able to use an object as validations property': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -278,7 +275,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const input = TestUtils.findRenderedComponentWithType(form, TestInput);
@@ -286,12 +284,10 @@ export default {
     test.equal(input.isValidValue('foo@bar'), false);
 
     test.done();
-
   },
 
   'should be able to pass complex values to a validation rule': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -301,7 +297,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
@@ -311,18 +308,18 @@ export default {
     test.equal(inputComponent.isValid(), false);
 
     test.done();
-
   },
 
   'should be able to run a function to validate': function (test) {
-
-    const TestForm = React.createClass({
-      customValidationA(values, value) {
+    class TestForm extends React.Component {
+      customValidationA = (values, value) => {
         return value === 'foo';
-      },
-      customValidationB(values, value) {
+      };
+
+      customValidationB = (values, value) => {
         return value === 'foo' && values.A === 'foo';
-      },
+      };
+
       render() {
         return (
           <Formsy.Form>
@@ -335,7 +332,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.scryRenderedComponentsWithType(form, TestInput);
@@ -347,12 +345,10 @@ export default {
     test.equal(inputComponent[1].isValid(), false);
 
     test.done();
-
   },
 
   'should not override error messages with error messages passed by form if passed eror messages is an empty object': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form validationErrors={{}}>
@@ -362,20 +358,19 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
     test.equal(inputComponent.getErrorMessage(), 'bar3');
 
     test.done();
-
   },
 
 
   'should override all error messages with error messages passed by form': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form validationErrors={{A: 'bar'}}>
@@ -385,19 +380,18 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
     test.equal(inputComponent.getErrorMessage(), 'bar');
 
     test.done();
-
   },
 
   'should override validation rules with required rules': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -415,19 +409,18 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
     test.equal(inputComponent.getErrorMessage(), 'bar3');
 
     test.done();
-
   },
 
   'should fall back to default error message when non exist in validationErrors map': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -442,19 +435,18 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
     test.equal(inputComponent.getErrorMessage(), 'bar');
 
     test.done();
-
   },
 
   'should not be valid if it is required and required rule is true': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -464,25 +456,23 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const inputComponent = TestUtils.findRenderedComponentWithType(form, TestInput);
     test.equal(inputComponent.isValid(), false);
 
     test.done();
-
   },
 
   'should handle objects and arrays as values': function (test) {
+    class TestForm extends React.Component {
+      state = {
+        foo: {foo: 'bar'},
+        bar: ['foo']
+      };
 
-    const TestForm = React.createClass({
-      getInitialState() {
-        return {
-          foo: {foo: 'bar'},
-          bar: ['foo']
-        };
-      },
       render() {
         return (
           <Formsy.Form>
@@ -491,7 +481,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     form.setState({
@@ -504,22 +495,20 @@ export default {
     test.deepEqual(inputs[1].getValue(), ['bar']);
 
     test.done();
-
   },
 
   'should handle isFormDisabled with dynamic inputs': function (test) {
+    class TestForm extends React.Component {
+      state = {
+        bool: true
+      };
 
-    const TestForm = React.createClass({
-      getInitialState() {
-        return {
-          bool: true
-        };
-      },
-      flip() {
+      flip = () => {
         this.setState({
           bool: !this.state.bool
         });
-      },
+      };
+
       render() {
         return (
           <Formsy.Form disabled={this.state.bool}>
@@ -530,7 +519,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     const input = TestUtils.findRenderedComponentWithType(form, TestInput);
@@ -539,15 +529,14 @@ export default {
     test.equal(input.isFormDisabled(), false);
 
     test.done();
-
   },
 
   'should allow for dot notation in name which maps to a deep object': function (test) {
-
-    const TestForm = React.createClass({
-      onSubmit(model) {
+    class TestForm extends React.Component {
+      onSubmit = model => {
         test.deepEqual(model, {foo: {bar: 'foo', test: 'test'}});
-      },
+      };
+
       render() {
         return (
           <Formsy.Form onSubmit={this.onSubmit}>
@@ -556,7 +545,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     test.expect(1);
@@ -565,15 +555,14 @@ export default {
     TestUtils.Simulate.submit(formEl);
 
     test.done();
-
   },
 
   'should allow for application/x-www-form-urlencoded syntax and convert to object': function (test) {
-
-    const TestForm = React.createClass({
-      onSubmit(model) {
+    class TestForm extends React.Component {
+      onSubmit = model => {
         test.deepEqual(model, {foo: ['foo', 'bar']});
-      },
+      };
+
       render() {
         return (
           <Formsy.Form onSubmit={this.onSubmit}>
@@ -582,7 +571,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
     test.expect(1);
@@ -591,7 +581,6 @@ export default {
     TestUtils.Simulate.submit(formEl);
 
     test.done();
-
   },
 
   'input should rendered once with PureRenderMixin': function (test) {
@@ -619,10 +608,11 @@ export default {
   },
 
   'should not render the form when passed `skipFormRender`': function(test) {
-    const TestForm = React.createClass({
-      onSubmit(model) {
+    class TestForm extends React.Component {
+      onSubmit = model => {
         test.deepEqual(model, {foo: ['foo', 'bar']});
-      },
+      };
+
       render() {
         return (
           <div>
@@ -633,7 +623,8 @@ export default {
           </div>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     test.equal(ReactDOM.findDOMNode(form).getElementsByTagName("form").length, 0);
 
