@@ -1,15 +1,33 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 
 import Formsy from './..';
 import TestInput from './utils/TestInput';
+import TestInputHoc from './utils/TestInputHoc';
 import immediate from './utils/immediate';
 import sinon from 'sinon';
 
 export default {
 
   'Setting up a form': {
+    'should expose the users DOM node through an innerRef prop': function (test) {
+      class TestForm extends React.Component {
+        render() {
+          return (
+            <Formsy.Form>
+              <TestInputHoc name="name" innerRef={(c) => { this.name = c; }} />
+            </Formsy.Form>
+          );
+        }
+      }
+
+      const form = TestUtils.renderIntoDocument(<TestForm/>);
+      const input = form.name;
+      test.equal(input.methodOnWrappedInstance('foo'), 'foo');
+
+      test.done();
+    },
 
     'should render a form into the document': function (test) {
 
@@ -30,9 +48,9 @@ export default {
     },
 
     'should allow for null/undefined children': function (test) {
-
       let model = null;
-      const TestForm = React.createClass({
+
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onSubmit={(formModel) => (model = formModel)}>
@@ -43,7 +61,7 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       immediate(() => {
@@ -51,25 +69,26 @@ export default {
         test.deepEqual(model, {name: 'foo'});
         test.done();
       });
-
     },
 
     'should allow for inputs being added dynamically': function (test) {
-
       const inputs = [];
       let forceUpdate = null;
       let model = null;
-      const TestForm = React.createClass({
+
+      class TestForm extends React.Component {
         componentWillMount() {
           forceUpdate = this.forceUpdate.bind(this);
-        },
+        }
+
         render() {
           return (
             <Formsy.Form onSubmit={(formModel) => (model = formModel)}>
               {inputs}
             </Formsy.Form>);
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       // Wait before adding the input
@@ -87,25 +106,26 @@ export default {
         });
 
       }, 10);
-
     },
 
     'should allow dynamically added inputs to update the form-model': function (test) {
-
       const inputs = [];
       let forceUpdate = null;
       let model = null;
-      const TestForm = React.createClass({
+
+      class TestForm extends React.Component {
         componentWillMount() {
           forceUpdate = this.forceUpdate.bind(this);
-        },
+        }
+
         render() {
           return (
             <Formsy.Form onSubmit={(formModel) => (model = formModel)}>
               {inputs}
             </Formsy.Form>);
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       // Wait before adding the input
@@ -125,18 +145,17 @@ export default {
         });
 
       });
-
     },
 
     'should allow a dynamically updated input to update the form-model': function (test) {
-
       let forceUpdate = null;
       let model = null;
 
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         componentWillMount() {
           forceUpdate = this.forceUpdate.bind(this);
-        },
+        }
+
         render() {
           const input = <TestInput name="test" value={this.props.value} />;
 
@@ -145,7 +164,8 @@ export default {
               {input}
             </Formsy.Form>);
         }
-      });
+      }
+
       let form = TestUtils.renderIntoDocument(<TestForm value="foo"/>);
 
       // Wait before changing the input
@@ -163,7 +183,6 @@ export default {
         });
 
       });
-
     }
 
   },
@@ -336,18 +355,17 @@ export default {
   },
 
   'should not trigger onChange when form is mounted': function (test) {
-
-
     const hasChanged = sinon.spy();
-    const TestForm = React.createClass({
+
+    class TestForm extends React.Component {
       render() {
         return <Formsy.Form onChange={hasChanged}></Formsy.Form>;
       }
-    });
+    }
+
     TestUtils.renderIntoDocument(<TestForm/>);
     test.equal(hasChanged.called, false);
     test.done();
-
   },
 
   'should trigger onChange once when form element is changed': function (test) {
@@ -365,19 +383,19 @@ export default {
   },
 
   'should trigger onChange once when new input is added to form': function (test) {
-
     const hasChanged = sinon.spy();
-    const TestForm = React.createClass({
-      getInitialState() {
-        return {
-          showInput: false
-        };
-      },
-      addInput() {
+
+    class TestForm extends React.Component {
+      state = {
+        showInput: false
+      };
+
+      addInput = () => {
         this.setState({
           showInput: true
         })
-      },
+      };
+
       render() {
         return (
           <Formsy.Form onChange={hasChanged}>
@@ -389,7 +407,7 @@ export default {
             }
           </Formsy.Form>);
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     form.addInput();
@@ -397,23 +415,22 @@ export default {
       test.equal(hasChanged.calledOnce, true);
       test.done();
     });
-
   },
 
   'Update a form': {
 
     'should allow elements to check if the form is disabled': function (test) {
+      class TestForm extends React.Component {
+        state = { disabled: true };
+        enableForm = () => { this.setState({ disabled: false }); };
 
-      const TestForm = React.createClass({
-        getInitialState() { return { disabled: true }; },
-        enableForm() { this.setState({ disabled: false }); },
         render() {
           return (
             <Formsy.Form disabled={this.state.disabled}>
               <TestInput name="foo"/>
             </Formsy.Form>);
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
@@ -424,23 +441,24 @@ export default {
         test.equal(input.isFormDisabled(), false);
         test.done();
       });
-
     },
 
     'should be possible to pass error state of elements by changing an errors attribute': function (test) {
+      class TestForm extends React.Component {
+        state = { validationErrors: { foo: 'bar' } };
 
-      const TestForm = React.createClass({
-        getInitialState() { return { validationErrors: { foo: 'bar' } }; },
-        onChange(values) {
+        onChange = values => {
             this.setState(values.foo ? { validationErrors: {} } : { validationErrors: {foo: 'bar'} });
-        },
+        };
+
         render() {
           return (
             <Formsy.Form onChange={this.onChange} validationErrors={this.state.validationErrors}>
               <TestInput name="foo"/>
             </Formsy.Form>);
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       // Wait for update
@@ -455,39 +473,39 @@ export default {
           test.done();
         });
       });
-
     },
 
     'should trigger an onValidSubmit when submitting a valid form': function (test) {
-
       let isCalled = sinon.spy();
-      const TestForm = React.createClass({
+
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onValidSubmit={isCalled}>
               <TestInput name="foo" validations="isEmail" value="foo@bar.com"/>
             </Formsy.Form>);
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const FoundForm = TestUtils.findRenderedComponentWithType(form, TestForm);
       TestUtils.Simulate.submit(ReactDOM.findDOMNode(FoundForm));
       test.equal(isCalled.called,true);
       test.done();
-
     },
 
     'should trigger an onInvalidSubmit when submitting an invalid form': function (test) {
-
       let isCalled = sinon.spy();
-      const TestForm = React.createClass({
+
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onInvalidSubmit={isCalled}>
               <TestInput name="foo" validations="isEmail" value="foo@bar"/>
             </Formsy.Form>);
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
 
       const FoundForm = TestUtils.findRenderedComponentWithType(form, TestForm);
@@ -495,7 +513,6 @@ export default {
       test.equal(isCalled.called, true);
 
       test.done();
-
     }
 
   },
@@ -503,9 +520,9 @@ export default {
   'value === false': {
 
     'should call onSubmit correctly': function (test) {
-
       const onSubmit = sinon.spy();
-      const TestForm = React.createClass({
+
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form onSubmit={onSubmit}>
@@ -514,29 +531,28 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       TestUtils.Simulate.submit(ReactDOM.findDOMNode(form));
       test.equal(onSubmit.calledWith({foo: false}), true);
       test.done();
-
     },
 
     'should allow dynamic changes to false': function (test) {
-
       const onSubmit = sinon.spy();
-      const TestForm = React.createClass({
-        getInitialState() {
-          return {
-            value: true
-          };
-        },
-        changeValue() {
+
+      class TestForm extends React.Component {
+        state = {
+          value: true
+        };
+
+        changeValue = () => {
           this.setState({
             value: false
           });
-        },
+        };
+
         render() {
           return (
             <Formsy.Form onSubmit={onSubmit}>
@@ -545,19 +561,17 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
 
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       form.changeValue();
       TestUtils.Simulate.submit(ReactDOM.findDOMNode(form));
       test.equal(onSubmit.calledWith({foo: false}), true);
       test.done();
-
     },
 
     'should say the form is submitted': function (test) {
-
-      const TestForm = React.createClass({
+      class TestForm extends React.Component {
         render() {
           return (
             <Formsy.Form>
@@ -566,29 +580,28 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
       test.equal(input.isFormSubmitted(), false);
       TestUtils.Simulate.submit(ReactDOM.findDOMNode(form));
       test.equal(input.isFormSubmitted(), true);
       test.done();
-
     },
 
     'should be able to reset the form to its pristine state': function (test) {
+      class TestForm extends React.Component {
+        state = {
+          value: true
+        };
 
-      const TestForm = React.createClass({
-        getInitialState() {
-          return {
-            value: true
-          };
-        },
-        changeValue() {
+        changeValue = () => {
           this.setState({
             value: false
           });
-        },
+        };
+
         render() {
           return (
             <Formsy.Form>
@@ -597,7 +610,8 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
       const formsyForm = TestUtils.findRenderedComponentWithType(form, Formsy.Form);
@@ -608,22 +622,20 @@ export default {
       test.equal(input.getValue(), true);
 
       test.done();
-
     },
 
     'should be able to reset the form using custom data': function (test) {
+      class TestForm extends React.Component {
+        state = {
+          value: true
+        };
 
-      const TestForm = React.createClass({
-        getInitialState() {
-          return {
-            value: true
-          };
-        },
-        changeValue() {
+        changeValue = () => {
           this.setState({
             value: false
           });
-        },
+        };
+
         render() {
           return (
             <Formsy.Form>
@@ -632,7 +644,8 @@ export default {
             </Formsy.Form>
           );
         }
-      });
+      }
+
       const form = TestUtils.renderIntoDocument(<TestForm/>);
       const input = TestUtils.findRenderedComponentWithType(form, TestInput);
       const formsyForm = TestUtils.findRenderedComponentWithType(form, Formsy.Form);
@@ -645,14 +658,12 @@ export default {
       });
       test.equal(input.getValue(), 'bar');
       test.done();
-
     }
 
   },
 
   'should be able to reset the form to empty values': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form>
@@ -661,7 +672,8 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
+
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const input = TestUtils.findRenderedComponentWithType(form, TestInput);
     const formsyForm = TestUtils.findRenderedComponentWithType(form, Formsy.Form);
@@ -671,7 +683,6 @@ export default {
     });
     test.equal(input.getValue(), '');
     test.done();
-
   },
 
   '.isChanged()': {
