@@ -1,5 +1,5 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 
 import Formsy from './..';
 import TestInput, {InputFactory} from './utils/TestInput';
@@ -111,8 +111,7 @@ export default {
   },
 
   'should provide invalidate callback on onValiSubmit': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form onValidSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
@@ -120,7 +119,7 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
 
@@ -129,12 +128,10 @@ export default {
     TestUtils.Simulate.submit(formEl);
     test.equal(input.isValid(), false);
     test.done();
-
   },
 
   'should provide invalidate callback on onInvalidSubmit': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form onInvalidSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
@@ -142,7 +139,7 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
@@ -151,12 +148,10 @@ export default {
     test.equal(input.getErrorMessage(), 'bar');
 
     test.done();
-
   },
 
   'should not invalidate inputs on external errors with preventExternalInvalidation prop': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form
@@ -166,7 +161,7 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
@@ -174,12 +169,10 @@ export default {
     TestUtils.Simulate.submit(formEl);
     test.equal(input.isValid(), true);
     test.done();
-
   },
 
   'should invalidate inputs on external errors without preventExternalInvalidation prop': function (test) {
-
-    const TestForm = React.createClass({
+    class TestForm extends React.Component {
       render() {
         return (
           <Formsy.Form onSubmit={(model, reset, invalidate) => invalidate({ foo: 'bar' })}>
@@ -187,7 +180,7 @@ export default {
           </Formsy.Form>
         );
       }
-    });
+    }
 
     const form = TestUtils.renderIntoDocument(<TestForm/>);
     const formEl = TestUtils.findRenderedDOMComponentWithTag(form, 'form');
@@ -195,7 +188,6 @@ export default {
     TestUtils.Simulate.submit(formEl);
     test.equal(input.isValid(), false);
     test.done();
-
   }
 
 };
